# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitHub API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/workouts": {
      "get": {
        "tags": [
          "workouts"
        ],
        "summary": "List Workouts Route",
        "operationId": "list_workouts_route_api_v1_workouts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "before_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "session_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Type"
            }
          },
          {
            "name": "partner_only",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Partner Only"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workouts"
        ],
        "summary": "Create Workout Route",
        "operationId": "create_workout_route_api_v1_workouts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workouts/parse-nl": {
      "post": {
        "tags": [
          "workouts"
        ],
        "summary": "Parse Workout Nl",
        "description": "Minimal NL workout parser \u2014 extracts title from free-form text.",
        "operationId": "parse_workout_nl_api_v1_workouts_parse_nl_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseNLRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseNLResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/workouts/{workout_id}": {
      "get": {
        "tags": [
          "workouts"
        ],
        "summary": "Get Workout Route",
        "operationId": "get_workout_route_api_v1_workouts__workout_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "workouts"
        ],
        "summary": "Patch Workout Route",
        "operationId": "patch_workout_route_api_v1_workouts__workout_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workout Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchWorkoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workouts"
        ],
        "summary": "Delete Workout Route",
        "operationId": "delete_workout_route_api_v1_workouts__workout_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workouts/{workout_id}/team-session": {
      "get": {
        "tags": [
          "workouts"
        ],
        "summary": "Get Workout Team Session Route",
        "operationId": "get_workout_team_session_route_api_v1_workouts__workout_id__team_session_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/wellness/checkin": {
      "post": {
        "tags": [
          "wellness"
        ],
        "summary": "Submit Checkin",
        "description": "Submit today's Hooper Index check-in (upserts \u2014 one per calendar day).",
        "operationId": "submit_checkin_api_v1_wellness_checkin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/wellness/checkin/today": {
      "get": {
        "tags": [
          "wellness"
        ],
        "summary": "Today Checkin",
        "description": "Return today's check-in, or submitted=false if not yet logged.",
        "operationId": "today_checkin_api_v1_wellness_checkin_today_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayCheckInResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckInRequest": {
        "properties": {
          "sleep": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Sleep"
          },
          "stress": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Stress"
          },
          "fatigue": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Fatigue"
          },
          "soreness": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Soreness"
          }
        },
        "type": "object",
        "required": [
          "sleep",
          "stress",
          "fatigue",
          "soreness"
        ],
        "title": "CheckInRequest"
      },
...
      "CreateWorkoutRequest": {
        "properties": {
          "performed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Performed At"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "bodyweight_kg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 600.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyweight Kg"
          },
          "session_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "workout_format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkoutFormat"
              },
              {
                "type": "null"
              }
            ]
          },
          "time_cap_s": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Cap S"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "session_rpe": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Rpe"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "is_tag": {
            "type": "boolean",
            "title": "Is Tag",
            "default": false
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CreateResultRequest"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "performed_at"
        ],
        "title": "CreateWorkoutRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParseNLRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 5000,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ParseNLRequest"
      },
      "ParseNLResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "title",
          "notes"
        ],
        "title": "ParseNLResponse"
      },
...
      "PatchWorkoutRequest": {
        "properties": {
          "performed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Performed At"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "bodyweight_kg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 600.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyweight Kg"
          },
          "session_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "workout_format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkoutFormat"
              },
              {
                "type": "null"
              }
            ]
          },
          "time_cap_s": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Cap S"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "session_rpe": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Rpe"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          }
        },
        "type": "object",
        "title": "PatchWorkoutRequest"
      },
...
      "SessionType": {
        "type": "string",
        "enum": [
          "strength",
          "metcon",
          "skill",
          "mixed",
          "rest",
          "deload",
          "active_recovery"
        ],
        "title": "SessionType"
      },
...
      "Workout": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "performed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Performed At"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "short_hash": {
            "type": "string",
            "title": "Short Hash"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "bodyweight_kg": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyweight Kg"
          },
          "session_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "workout_format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkoutFormat"
              },
              {
                "type": "null"
              }
            ]
          },
          "time_cap_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Cap S"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "session_rpe": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Rpe"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "perceived_load_au": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perceived Load Au"
          },
          "volume_load_kg": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Load Kg"
          },
          "avg_hr": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Hr"
          },
          "max_hr": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Hr"
          },
          "trimp_au": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trimp Au"
          },
          "is_tag": {
            "type": "boolean",
            "title": "Is Tag",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Result"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "performed_at",
          "title",
          "short_hash",
          "notes",
          "bodyweight_kg",
          "session_type",
          "workout_format",
          "time_cap_s",
          "location",
          "session_rpe",
          "duration_s",
          "perceived_load_au",
          "volume_load_kg",
          "avg_hr",
          "max_hr",
          "trimp_au",
          "created_at",
          "updated_at"
        ],
        "title": "Workout"
      },
...
      "WorkoutListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkoutSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "WorkoutListResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}